+++ GitHub4Jive-Addon/services/github/backend/gitHubController.js
@@ -40,7 +40,7 @@ var JiveAuth = require(libDir + "github4jive/JiveOauth");
 exports.gitHubWebHookPortal = function (req, res) {
     var event = req.headers["x-github-event"];
     var data = req.body;
-    jive.logger.info("Received GitHub data: " + event)
+    jive.logger.info("Received GitHub data: " + event);
     gitHubFacade.notifyNewGitHubHookInfo(event, data);
     res.writeHead(202);
     res.end();
