+++ GitHub4Jive-Addon/services/github/backend/gitHubOAuthController.js
