+++ GitHub4Jive-Addon/services/jive/backend/jiveOAuthController.js
@@ -35,7 +35,7 @@ myOauth.fetchOAuth2Conf = function() {
 };
 
 myOauth.oauth2SuccessCallback = function( state, originServerAccessTokenResponse, callback ) {
-  console.log("************************* OAUTH CALLBACK");
+    console.log("************************* OAUTH CALLBACK");
     console.log('************************* State', state);
     console.log('************************* originServerAccessTokenResponse', originServerAccessTokenResponse);
 
@@ -45,8 +45,7 @@ myOauth.oauth2SuccessCallback = function( state, originServerAccessTokenResponse
     toStore.jive = originServerAccessTokenResponse.entity;
     toStore.jive.userID =  state['viewerID'];
     placeStore.save(placeRef, toStore).then( function() {
-          console.log('************************* doing callback');
-
+        console.log('************************* doing callback');
         callback({'ticket': state['viewerID'] });
     });
 };
